moved getBounds tests to MemoryDataStoreBoundsTest
