+++ modules/library/main/src/test/java/org/geotools/data/memory/MemoryDataStoreBoundsTest.java
@@ -17,11 +17,17 @@
 package org.geotools.data.memory;
 
 import org.geotools.data.DataTestCase;
+import org.geotools.data.DataUtilities;
 import org.geotools.data.DefaultQuery;
 import org.geotools.data.Query;
+import org.geotools.data.simple.SimpleFeatureSource;
+import org.geotools.feature.simple.SimpleFeatureBuilder;
+import org.opengis.feature.simple.SimpleFeature;
+import org.opengis.feature.simple.SimpleFeatureType;
 
 /**
  * @author Frank Gasdorf, fgdrf@users.sourceforge.net
+ * @author Sebastian Graca, ISPiK S.A.
  *
  *
  *
@@ -59,4 +65,27 @@ public class MemoryDataStoreBoundsTest extends DataTestCase {
         assertEquals(roadFeatures[1].getBounds(), data.getBounds(query));
     }
 
+    public void testGetBoundsSupportsFeaturesWithoutGeometry() throws Exception {
+        SimpleFeatureBuilder featureBuilder = new SimpleFeatureBuilder(data.getSchema("road"));
+        featureBuilder.init(roadFeatures[0]);
+        featureBuilder.set("geom", null);
+        SimpleFeature feature = featureBuilder.buildFeature("road.rd0");
+        data.addFeature(feature);
+
+        SimpleFeatureSource road = data.getFeatureSource("road");
+        assertEquals(roadBounds, road.getBounds(Query.ALL));
+    }
+
+    public void testGetBoundsSupportsEmptyBounds() throws Exception {
+        SimpleFeatureType type = DataUtilities.createType(getName() + ".test",
+                "id:0,geom:LineString,name:String");
+        SimpleFeature[] features = new SimpleFeature[3];
+        features[0] = SimpleFeatureBuilder.build(type, new Object[] {1, null, "r1"}, "test.f1");
+        features[1] = SimpleFeatureBuilder.build(type, new Object[] {2, null, "r2"}, "test.f2");
+        features[2] = SimpleFeatureBuilder.build(type, new Object[] {3, null, "r3"}, "test.f3");
+        data.addFeatures(features);
+
+        SimpleFeatureSource featureSource = data.getFeatureSource("test");
+        assertTrue(featureSource.getBounds(Query.ALL).isEmpty());
+    }
 }
